fix: /tx trace tab loading
`traceQueryOptions()` now calls the RPC directly rather than via the API,
which was failing during SSR because of `__BASE_URL__`.
